% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{constructhelpers}
\alias{constructhelpers}
\title{constructhelpers}
\description{
Simpler/more efficient implementations of some bits of gbradburd's
conStruct. Specifically, a one-stop-shop, fully parallelised
cross-validation workflow.
}
\author{
Kevin Murray
}
