% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xval.R
\name{csh.x.validation}
\alias{csh.x.validation}
\title{csh.x.validation}
\usage{
csh.x.validation(
  freqs = NULL,
  data.partitions = NULL,
  geoDist,
  coords,
  prefix,
  K,
  n.reps = 8,
  n.iter = 10000,
  train.prop = 0.9,
  make.figs = FALSE,
  save.files = FALSE,
  ...
)
}
\arguments{
\item{freqs}{Per-locality allele frequencies, see conStruct::conStruct()}

\item{data.partitions}{Pre-computed data partitions, see conStruct::x.validation()}

\item{geoDist}{Geographic distance matrix, see conStruct::conStruct()}

\item{prefix}{File prefix for any output files, see conStruct::x.validation()}

\item{K}{Vector of layers/ancestral pops. (e.g. 1:6)}

\item{n.reps}{Number of replicate runs/mcmc chains}

\item{n.iter}{Number of MCMC interations}

\item{train.prop}{Proportion of SNPs to use for training vs test. typically 0.8-0.9}

\item{make.figs}{Save plots to disk}

\item{save.files}{Save intermediates to disk}

\item{coods}{X/Y or Long/Lat coordiate matrix, see conStruct::conStruct()}
}
\value{
tibble with one row per MCMC run (so length(K) * n.reps * 2)
}
\description{
csh.x.validation
}
